--- Debug/CPP/src/main.su
@@ -1,5 +1,0 @@
-../CPP/src/main.cpp:24:6:void maincpp()	8	static
-../CPP/src/main.cpp:77:6:void HAL_GPIO_EXTI_Rising_Callback(uint16_t)	24	static
-../CPP/src/main.cpp:99:6:void HAL_GPIO_EXTI_Falling_Callback(uint16_t)	16	static
-../CPP/src/main.cpp:106:1:void __static_initialization_and_destruction_0()	8	static
-../CPP/src/main.cpp:106:1:cpp)	8	static
