+++ Debug/makefile
@@ -12,7 +12,7 @@ RM := rm -rf
 -include Drivers/STM32G0xx_HAL_Driver/Src/subdir.mk
 -include Core/Startup/subdir.mk
 -include Core/Src/subdir.mk
--include CPP/src/subdir.mk
+-include CPP/subdir.mk
 -include objects.mk
 
 ifneq ($(MAKECMDGOALS),clean)
@@ -59,34 +59,34 @@ $(wildcard ../makefile.init) \
 $(wildcard ../makefile.targets) \
 
 
-BUILD_ARTIFACT_NAME := INTERRUPTS
+BUILD_ARTIFACT_NAME := INTERRUPTS PT2
 BUILD_ARTIFACT_EXTENSION := elf
 BUILD_ARTIFACT_PREFIX :=
 BUILD_ARTIFACT := $(BUILD_ARTIFACT_PREFIX)$(BUILD_ARTIFACT_NAME)$(if $(BUILD_ARTIFACT_EXTENSION),.$(BUILD_ARTIFACT_EXTENSION),)
 
 # Add inputs and outputs from these tool invocations to the build variables 
 EXECUTABLES += \
-INTERRUPTS.elf \
+INTERRUPTS\ PT2.elf \
 
 MAP_FILES += \
-INTERRUPTS.map \
+INTERRUPTS\ PT2.map \
 
 SIZE_OUTPUT += \
 default.size.stdout \
 
 OBJDUMP_LIST += \
-INTERRUPTS.list \
+INTERRUPTS\ PT2.list \
 
 
 # All Target
 all: main-build
 
 # Main-build Target
-main-build: INTERRUPTS.elf secondary-outputs
+main-build: INTERRUPTS\ PT2.elf secondary-outputs
 
 # Tool invocations
-INTERRUPTS.elf INTERRUPTS.map: $(OBJS) $(USER_OBJS) C:\Users\Gesse\STM32CubeIDE\workspace_1.18.1\INTERRUPTS\STM32G071RBTX_FLASH.ld makefile objects.list $(OPTIONAL_TOOL_DEPS)
-	arm-none-eabi-g++ -o "INTERRUPTS.elf" @"objects.list" $(USER_OBJS) $(LIBS) -mcpu=cortex-m0plus -T"C:\Users\Gesse\STM32CubeIDE\workspace_1.18.1\INTERRUPTS\STM32G071RBTX_FLASH.ld" --specs=nosys.specs -Wl,-Map="INTERRUPTS.map" -Wl,--gc-sections -static --specs=nano.specs -mfloat-abi=soft -mthumb -Wl,--start-group -lc -lm -lstdc++ -lsupc++ -Wl,--end-group
+INTERRUPTS\ PT2.elf INTERRUPTS\ PT2.map: $(OBJS) $(USER_OBJS) C:\Users\Gesse\STM32CubeIDE\workspace_1.18.1\INTERRUPTS\ PT2\STM32G071RBTX_FLASH.ld makefile objects.list $(OPTIONAL_TOOL_DEPS)
+	arm-none-eabi-g++ -o "INTERRUPTS PT2.elf" @"objects.list" $(USER_OBJS) $(LIBS) -mcpu=cortex-m0plus -T"C:\Users\Gesse\STM32CubeIDE\workspace_1.18.1\INTERRUPTS PT2\STM32G071RBTX_FLASH.ld" --specs=nosys.specs -Wl,-Map="INTERRUPTS PT2.map" -Wl,--gc-sections -static --specs=nano.specs -mfloat-abi=soft -mthumb -Wl,--start-group -lc -lm -lstdc++ -lsupc++ -Wl,--end-group
 	@echo 'Finished building target: $@'
 	@echo ' '
 
@@ -95,14 +95,14 @@ default.size.stdout: $(EXECUTABLES) makefile objects.list $(OPTIONAL_TOOL_DEPS)
 	@echo 'Finished building: $@'
 	@echo ' '
 
-INTERRUPTS.list: $(EXECUTABLES) makefile objects.list $(OPTIONAL_TOOL_DEPS)
-	arm-none-eabi-objdump -h -S $(EXECUTABLES) > "INTERRUPTS.list"
+INTERRUPTS\ PT2.list: $(EXECUTABLES) makefile objects.list $(OPTIONAL_TOOL_DEPS)
+	arm-none-eabi-objdump -h -S $(EXECUTABLES) > "INTERRUPTS PT2.list"
 	@echo 'Finished building: $@'
 	@echo ' '
 
 # Other Targets
 clean:
-	-$(RM) INTERRUPTS.elf INTERRUPTS.list INTERRUPTS.map default.size.stdout
+	-$(RM) INTERRUPTS\ PT2.elf INTERRUPTS\ PT2.list INTERRUPTS\ PT2.map INTERRUPTS\\ PT2.elf default.size.stdout
 	-@echo ' '
 
 secondary-outputs: $(SIZE_OUTPUT) $(OBJDUMP_LIST)
