+++ INTERRUPTS PT2.ioc
@@ -3,7 +3,6 @@ CAD.formats=
 CAD.pinconfig=
 CAD.provider=
 File.Version=6
-GPIO.groupedBy=Group By Peripherals
 KeepUserPlacement=false
 Mcu.CPN=STM32G071RBT6
 Mcu.Family=STM32G0
@@ -104,10 +103,10 @@ ProjectManager.LastFirmware=true
 ProjectManager.LibraryCopy=1
 ProjectManager.MainLocation=Core/Src
 ProjectManager.NoMain=false
-ProjectManager.PreviousToolchain=STM32CubeIDE
+ProjectManager.PreviousToolchain=
 ProjectManager.ProjectBuild=false
-ProjectManager.ProjectFileName=INTERRUPTS.ioc
-ProjectManager.ProjectName=INTERRUPTS
+ProjectManager.ProjectFileName=INTERRUPTS PT2.ioc
+ProjectManager.ProjectName=INTERRUPTS PT2
 ProjectManager.ProjectStructure=
 ProjectManager.RegisterCallBack=
 ProjectManager.StackSize=0x400
